+++ vite.config.demo.ts
@@ -1,5 +1,4 @@
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
 
 export default defineConfig({
   base: '/node-data-components/',
@@ -12,7 +11,4 @@ export default defineConfig({
       treeshake: false,
     },
   },
-  plugins: [
-    react(),
-  ],
 });